Split out `rest_pat_in_fully_bound_struct`
